Fixed some bug
