+++ src/Components/Clusters/CreateCluster.js
@@ -9,7 +9,6 @@ export function CreateCluster (props) {
 	const classes = commonStyles();
 	const VMGroups = useSelector(state => state.vm_group);
 	const clusters = useSelector(state => state.clusters);
-	const [kubernetesVersions, setKubernetesVersions] = useState([]);
 	const [clusterName, setClusterName] = useState('');
 	const [kubernetesVersionId, setkubernetesVersionId] = useState('');
 	const [selectedVMGroup, setSelectedVMGroup] = useState('');
@@ -39,15 +38,8 @@ export function CreateCluster (props) {
 		})
 		.then(response => response.json())
 	}
-	const getKubernetesVersions = () => {
-		fetch(`${serverURL}/kubernetes/version/list`)
-		.then(response => response.json())
-		.then(data => data.kubernetes_version_list)
-		.then(kubernetes_version_list => setKubernetesVersions(kubernetes_version_list))
-		.then(console.log(getKubernetesVersions))
-	}
 
-	useEffect(getKubernetesVersions, []);
+	useEffect(props.getKubernetesVersions, []);
 	useEffect(props.refreshVMGroupData, []);
 	useEffect(props.refreshClustersData, []);
 	useEffect(() => {
@@ -96,7 +88,7 @@ export function CreateCluster (props) {
 				variant='outlined'
 				margin="dense"
 			>
-				{kubernetesVersions.map((version, i) => (
+				{props.kubernetesVersions.map((version, i) => (
 					<MenuItem key={i} value={version.id}>
 						{version.version}
 					</MenuItem>
